Improving YouTube app load speed
When there are larger videos to buffer, it takes forever for a video to load, but at the same time, we cannot start to play something at least without buffering a portion of it. So, we have picked a sensible mid point of 15%, which means there should be sufficient content for the video to be played while it is being buffered further. This will improve the load time of Content with many videos.